+++ src/lib/components/SheetToAddPayments.svelte
@@ -58,31 +58,31 @@
 			<div class="flex gap-4">
 				<div class="grid gap-2 w-full">
 					<Label for="name">Nombres</Label>
-					<Input type="text" id="name" name="name" placeholder="Pablito" required />
+					<Input class="text-lg" type="text" id="name" name="name"  required />
 				</div>
 				<div class="grid gap-2 w-full">
 					<Label for="rut">RUT</Label>
-					<Input type="text" id="rut" name="rut" placeholder="1111111-0" required />
+					<Input class="text-lg" type="text" id="rut" name="rut" />
 				</div>
 			</div>
 			<div class="flex gap-4">
 				<div class="grid gap-2 w-full">
 					<Label for="email">Email</Label>
-					<Input type="email" id="email" name="email" placeholder="pablito@5lc.cl" required />
+					<Input class="text-lg"  type="email" id="email" name="email" />
 				</div>
 				<div class="grid gap-2 w-full">
 					<Label for="phone">Teléfono</Label>
-					<Input id="phone" name="phone" type="text" placeholder="+56991291468" required />
+					<Input class="text-lg" id="phone" name="phone" type="text"  />
 				</div>
 			</div>
 			<div class="flex gap-4">
 				<div class="grid gap-2 w-full">
 					<Label for="ticketAmount">Entradas</Label>
-					<Input type="number" name="ticketAmount" min="0" max="10" id="ticketAmount" required />
+					<Input class="text-lg" type="number" name="ticketAmount" min="0" max="10" id="ticketAmount" value="1" required />
 				</div>
 				<div class="grid gap-2 w-full">
 					<Label for="price">Precio</Label>
-					<Input type="number" name="price" min="0" id="price" required />
+					<Input class="text-lg" type="number" name="price" min="0" id="price" required value="0" />
 				</div>
 			</div>
 			{#if $page.data.eventFromSanityStudio?.sell_type === 'ubication'}
@@ -94,7 +94,7 @@
 							type="radio"
 							id="general_tickets"
 							name="ticketType"
-							class="radio checked:bg-primary"
+							class="radio checked:bg-primary text-lg"
 							checked="checked"
 							value="general_tickets"
 						/>
@@ -105,7 +105,7 @@
 							type="radio"
 							name="ticketType"
 							id="ringside_tickets"
-							class="radio checked:bg-primary"
+							class="radio checked:bg-primary text-lg"
 							value="ringside_tickets"
 						/>
 						<span class="label-text mt-2">Ringisde</span>